Moved common fading code from Smoke, Bezele and Bubbles into FadingWindowController.

+++ Display Plugins/MusicVideo/GrowlMusicVideoWindowController.h
@@ -16,7 +16,7 @@
 	SEL				_action;
 	id				_target;
 	id				_representedObject;
-	short			_displayTime;
+	double			_displayTime;
 	float			topLeftPosition;
 	float			timerInterval;
 	float			fadeIncrement;
